Update media handling to new file references
CONNECT-25

+++ src/Receiver/MediaReceiver.php
@@ -5,14 +5,13 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Receiver;
 
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\File\FileReferenceContract;
 use Heptacom\HeptaConnect\Dataset\Ecommerce\Media\Media;
+use Heptacom\HeptaConnect\Portal\Base\File\FileReferenceResolverContract;
 use Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiveContextInterface;
 use Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiverContract;
-use Heptacom\HeptaConnect\Portal\Base\Serialization\Contract\DenormalizerInterface;
-use Heptacom\HeptaConnect\Portal\Base\Serialization\Contract\NormalizationRegistryContract;
 use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\DalAccess;
 use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\PrimaryKeyGenerator;
-use Psr\Http\Message\StreamInterface;
 use Shopware\Core\Content\Media\File\FileSaver;
 use Shopware\Core\Content\Media\File\MediaFile;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
@@ -24,17 +23,17 @@ class MediaReceiver extends ReceiverContract
 {
     private DalAccess $dal;
 
-    private NormalizationRegistryContract $normalizationRegistry;
+    private FileReferenceResolverContract $fileReferenceResolver;
 
     private FileSaver $fileSaver;
 
     public function __construct(
         DalAccess $dal,
-        NormalizationRegistryContract $normalizationRegistry,
+        FileReferenceResolverContract $fileReferenceResolver,
         FileSaver $fileSaver
     ) {
         $this->dal = $dal;
-        $this->normalizationRegistry = $normalizationRegistry;
+        $this->fileReferenceResolver = $fileReferenceResolver;
         $this->fileSaver = $fileSaver;
     }
 
@@ -51,13 +50,8 @@ class MediaReceiver extends ReceiverContract
         ReceiveContextInterface $context
     ): void {
         $swContext = $this->dal->getContext();
-        $source = $this->getMediaStream($entity);
         $mediaRepository = $this->dal->repository('media');
         $mediaFolderRepository = $this->dal->repository('media_folder');
-
-        if (!$source instanceof StreamInterface) {
-            throw new \RuntimeException('Could not unpack media stream');
-        }
 
         if (!$this->dal->idExists('media', $entity->getPrimaryKey())) {
             $criteria = new Criteria();
@@ -79,50 +73,26 @@ class MediaReceiver extends ReceiverContract
             'title' => $entity->getTitle()->getFallback(),
         ]], $swContext);
 
-        $fileName = \tempnam(\sys_get_temp_dir(), 'local-media-receiver');
+        $fileReference = $entity->getFile();
 
-        $this->fileSaver->persistFileToMedia(
-            self::prepareMediaFile($fileName, $entity->getMimeType(), $source),
-            $entity->getPrimaryKey(),
-            $entity->getPrimaryKey(),
-            $swContext
-        );
+        if ($fileReference instanceof FileReferenceContract) {
+            $file = $this->fileReferenceResolver->resolve($fileReference);
+            $fileName = \tempnam(\sys_get_temp_dir(), 'local-media-receiver');
 
-        if (@\is_file($fileName)) {
-            @\unlink($fileName);
+            try {
+                \file_put_contents($fileName, $file->getContents());
+                $swMediaFile = new MediaFile(
+                    $fileName,
+                    $entity->getMimeType(),
+                    MimeTypes::getDefault()->getExtensions($entity->getMimeType())[0] ?? 'bin',
+                    \filesize($fileName)
+                );
+                $this->fileSaver->persistFileToMedia($swMediaFile, $entity->getPrimaryKey(), $entity->getPrimaryKey(), $swContext);
+            } finally {
+                if (@\is_file($fileName)) {
+                    @\unlink($fileName);
+                }
+            }
         }
-    }
-
-    private static function prepareMediaFile(string $fileName, string $mimeType, StreamInterface $source): MediaFile
-    {
-        $sourceStream = $source->detach();
-        $file = \fopen($fileName, 'wb+');
-        \stream_copy_to_stream($sourceStream, $file);
-        \fclose($file);
-        \fclose($sourceStream);
-
-        return new MediaFile(
-            $fileName,
-            $mimeType,
-            MimeTypes::getDefault()->getExtensions($mimeType)[0] ?? 'bin',
-            \filesize($fileName)
-        );
-    }
-
-    private function getMediaStream(Media $media): ?StreamInterface
-    {
-        $denormalizer = $this->normalizationRegistry->getDenormalizer('stream');
-
-        if (!$denormalizer instanceof DenormalizerInterface) {
-            return null;
-        }
-
-        $stream = $denormalizer->denormalize($media->getNormalizedStream(), 'stream');
-
-        if (!$stream instanceof StreamInterface) {
-            return null;
-        }
-
-        return $stream;
     }
 }
